Reorganize modules and tests
This is a more "vertical" organization strategy, with modules grouped by
theme instead of by type.

Also, I'm experimenting with defining the state changes mapping in the
Sudoku object instead of in a separate file, with the idea that this
results in fewer modules needing to change for simple modifications, and
it reduces the distance between coupled pieces (specifically, trivial
action creators and trivial state changes).